@@ -13,12 +13,9 @@ import java.util.stream.IntStream;
 public class TimeApi {
 
     public static List<Date> getDatesBetweenUsingJava7(Date startDate, Date endDate) {
-        List<Date> datesInRange = new ArrayList<Date>();
-        Calendar calendar = new GregorianCalendar();
-        calendar.setTime(startDate);
-
-        Calendar endCalendar = new GregorianCalendar();
-        endCalendar.setTime(endDate);
+        List<Date> datesInRange = new ArrayList<>();
+        Calendar calendar = getCalendarWithoutTime(startDate);
+        Calendar endCalendar = getCalendarWithoutTime(endDate);
 
         while (calendar.before(endCalendar)) {
             Date result = calendar.getTime();
@@ -38,6 +35,17 @@ public class TimeApi {
 
     public static List<LocalDate> getDatesBetweenUsingJava9(LocalDate startDate, LocalDate endDate) {
         return startDate.datesUntil(endDate).collect(Collectors.toList());
+    }
+
+    private static Calendar getCalendarWithoutTime(Date date) {
+        Calendar calendar = new GregorianCalendar();
+        calendar.setTime(date);
+        calendar.set(Calendar.HOUR, 0);
+        calendar.set(Calendar.HOUR_OF_DAY, 0);
+        calendar.set(Calendar.MINUTE, 0);
+        calendar.set(Calendar.SECOND, 0);
+        calendar.set(Calendar.MILLISECOND, 0);
+        return calendar;
     }
 
 }
