@@ -51,7 +51,7 @@ public class LogsResourceIntTest {
     public void changeLogs() throws Exception {
         LoggerVM logger = new LoggerVM();
         logger.setLevel("INFO");
-        logger.setName("ROOT");
+        logger.setName("some.test.logger");
 
         restLogsMockMvc.perform(put("/management/logs")
             .contentType(TestUtil.APPLICATION_JSON_UTF8)
