@@ -19,7 +19,7 @@ public class MemoryAppender extends ListAppender<ILoggingEvent> {
 
     public boolean contains(String string, Level level) {
         return this.list.stream()
-          .anyMatch(event -> event.getMessage().toString().contains(string) 
+          .anyMatch(event -> event.toString().contains(string) 
             && event.getLevel().equals(level));
     }
 
@@ -30,13 +30,13 @@ public class MemoryAppender extends ListAppender<ILoggingEvent> {
 
     public List<ILoggingEvent> search(String string) {
         return this.list.stream()
-          .filter(event -> event.getMessage().toString().contains(string))
+          .filter(event -> event.toString().contains(string))
           .collect(Collectors.toList());
     }
 
     public List<ILoggingEvent> search(String string, Level level) {
         return this.list.stream()
-          .filter(event -> event.getMessage().toString().contains(string) 
+          .filter(event -> event.toString().contains(string) 
             && event.getLevel().equals(level))
           .collect(Collectors.toList());
     }
